refactor(checker): split binary computation helpers
